RSTT.Syntax.ErrM

Description

BNF Converter: Error Monad.

Module for backwards compatibility.

The generated parser now uses Either String as error monad. This module defines a type synonym Err and pattern synonyms Bad and Ok for Left and Right.

Documentation

type Err = Either String

Error monad with String error messages.

pattern Bad :: a -> Either a b

pattern Ok :: forall {a} {b}. b -> Either a b

Orphan instances

MonadFail Err

Instance details

Methods

fail :: String -> Err a

Alternative Err

Instance details

Methods

empty :: Err a

(<|>) :: Err a -> Err a -> Err a

some :: Err a -> Err [a]

many :: Err a -> Err [a]

MonadPlus Err

Instance details

Methods

mzero :: Err a

mplus :: Err a -> Err a -> Err a