RSTT.Syntax.Layout

Documentation

topLayout :: Bool

layoutWords :: [String]

layoutStopWords :: [String]

layoutOpen :: String

layoutClose :: String

layoutSep :: String

resolveLayout

Arguments

:: Bool
Whether to use top-level layout.

-> [Token]

-> [Token]

Replace layout syntax with explicit layout tokens.

data Block

Constructors

Implicit Int
An implicit layout block with its start column.

Explicit

indentation :: [Block] -> Int

Get current indentation. 0 if we are in an explicit block.

isImplicit :: Block -> Bool

Check if s block is implicit.

type Position = Posn

addTokens

Arguments

:: Position
Position of the first new token.

-> [String]
Token symbols.

-> [Token]
The rest of the tokens. These will have their positions updated to make room for the new tokens .

-> [Token]

Insert a number of tokens at the begninning of a list of tokens.

addToken

Arguments

:: Position
Position of the new token.

-> String
Symbol in the new token.

-> [Token]
The rest of the tokens. These will have their positions updated to make room for the new token.

-> [Token]

Insert a new symbol token at the begninning of a list of tokens.

afterPrev :: Maybe Token -> Position

Get the position immediately to the right of the given token. If no token is given, gets the first position in the file.

nextPos :: Token -> Position

Get the position immediately to the right of the given token.

incrGlobal

Arguments

:: Position
If the token is on the same line as this position, update the column position.

-> Int
Number of characters to add to the position.

-> Token

-> Token

Add to the global and column positions of a token. The column position is only changed if the token is on the same line as the given position.

sToken :: Position -> String -> Token

Create a symbol token.

position :: Token -> Position

Get the position of a token.

line :: Token -> Int

Get the line number of a token.

column :: Token -> Int

Get the column number of a token.

isTokenIn :: [String] -> Token -> Bool

Check if a token is one of the given symbols.

isLayout :: Token -> Bool

Check if a word is a layout start token.

isStop :: Token -> Bool

Check if a token is a layout stop token.

isLayoutOpen :: Token -> Bool

Check if a token is the layout open token.

isLayoutClose :: Token -> Bool

Check if a token is the layout close token.

tokenLength :: Token -> Int

Get the number of characters in the token.