RSTT.Tope

Documentation

data Tope

A tope expression.

Constructors

TopeTop
Top tope: \(\top\).

TopeBottom
Bottom tope: \(\bot\).

TopeAnd Tope Tope
AND tope: \(\phi \land \psi\).

TopeOr Tope Tope
OR tope: \(\phi \lor \psi\).

TopeEQ Point Point
Point equality tope: \(t \equiv s\).

TopeVar Var
A tope variable.

TopeImplies Tope Tope
IMPLIES tope: \(\phi \Rightarrow \psi\). NOTE: this tope is added to internalise known tope judgements and simplify proof search.

TopeCon Label [Point]
User-defined tope constructor (fully applied to some points).

data Shape

A shape is a tope in a cube context (with just one cube point variable): \(\{t : I \mid \phi \}\).

Constructors

Shape

Fields

shapePoint :: (Var, Cube)

shapeTope :: Tope

type TopeContext = [Tope]

A tope context consists of a set of topes.

Pretty-printing

ppTopeContext :: TopeContext -> String

Pretty-print a TopeContext.

ppShape :: Shape -> String

Pretty-print a Shape.

ppTope :: Tope -> String

Pretty-print a Tope.

ppTopePrec :: Int -> Tope -> String

Pretty-print a Tope expression in a given precedence environment.