RSTT.Tope.Proof

Documentation

>>> :set -XOverloadedStrings
>>> import RSTT.Tope

Tope sequent

data Sequent

Tope sequent \( \Xi \mid \Phi \vdash \psi \).

Constructors

Sequent

Fields

sequentCubeContext :: CubeContext

Cube context \(\Xi\).

sequentTopeContext :: TopeContext

Tope context \(\Phi\).

sequentTope :: Tope

A tope \(\psi\) on the right hand side.

ppSequent :: Sequent -> String

Pretty-print a Sequent.

>>> points = [("t", "I"), ("s", "J")]
>>> topes = [("φ" `TopeOr` "ζ") `TopeAnd` ("ψ" `TopeOr` "χ"), ("t" `TopeEQ` "s")]
>>> putStrLn (ppSequent (Sequent points topes ("s" `TopeEQ` "t")))
t : I, s : J | (φ ∨ ζ) ∧ (ψ ∨ χ), t ≡ s ⊢ s ≡ t

Rules

type RuleName = String

A name of a (tope inference) rule.

type Rules = RulesM (RuleName, [Sequent])

A collection of tope inference rules. Intuitively, Rules is a non-deterministic function from Sequent to (RuleName, [Sequent]).

newtype RulesM a

A collection of generalized tope inference rules. Intuitively, RulesM a is a non-deterministic function from Sequent to a.

Constructors

RulesM

Fields

applyRulesM :: ReaderT Sequent Logic a

Sets of rules

rulesLJ :: Rules

Rules for intuitionistic tope logic:

axiom rule

\(\land\)-rules (leftAnd and rightAnd)

\(\lor\)-rules (leftOr, rightOrL, and rightOrR)

\(\Rightarrow\)-rules (leftImplies, rightImplies)

rulesEQ :: Rules

Rules for tope equality:

reflexivity (reflEQ)

symmetry (symEQ)

transitivity (transEQ)

rulesLJE :: Rules

Rules for intuitionistic tope logic with equality:

intuitionistic rules rulesLJ

equality tope rules rulesEQ

rulesLEQ :: Rules

Rules for inequality tope (not built-in).

reflexivity (reflLEQ)

antisymmetry (antisymLEQ)

transitivity (transLEQ)

excluded middle (lemLEQ)

infimum (zeroLEQ)

supremum (oneLEQ)

distinctness of 0 and 1 (distinctLEQ)

Intuitionistic logic rules (LJ)

ruleTop :: Rules

ruleBottom :: Rules

leftAnd :: Rules

leftOr :: Rules

leftImplies :: Rules

rightImplies :: Rules

rightAnd :: Rules

rightOrL :: Rules

rightOrR :: Rules

axiom :: Rules

Equality tope \((\equiv)\) rules

reflEQ :: Rules

Transitivity rule for equality tope (see TopeEQ).

———————————
 ⋅ ⊢ x ≡ x

transEQ :: Rules

Transitivity rule for equality tope (see TopeEQ).

     x ≡ y, y ≡ z ⊢ φ
—————————————————————————
 x ≡ y, y ≡ z, x ≡ z ⊢ φ

symEQ :: Rules

Symmetry rule for equality tope (see TopeEQ).

    x ≡ y ⊢ φ
——————————————————
 x ≡ y, y ≡ x ⊢ φ

Inequality tope \((\leq)\) rules

reflLEQ :: Rules

transLEQ :: Rules

antisymLEQ :: Rules

lemLEQ :: Rules

zeroLEQ :: Rules

oneLEQ :: Rules

distinctLEQ :: Rules

Helpers

selectN :: Alternative f => Int -> [a] -> f ([a], [a])

Nondeterministically select exactly \(N\) elements from a given list, also returning the remaining part of the list.

selectOne :: (Alternative f, MonadFail f) => [a] -> f (a, [a])

Nondeterministically select exactly one element from a given list, also returning the remaining part of the list.

choose :: Alternative f => [a] -> f a

Nondeterministically select exactly one element from a given list.

Proof tree

type Proof = ProofTree Void

A complete proof is a proof tree without any unfinished parts.

data ProofTree a

A proof tree parametrized by the type of leaves (unfinished proofs).

Constructors

Leaf a
A leaf represents an unfinished part of a proof.

Node Sequent RuleName [ProofTree a]
A node in a proof tree corresponds to a single rule application.

close :: Alternative f => ProofTree a -> f Proof

Convert a proof tree into a complete proof if possible.

ASCII rendering

ppProof :: Proof -> String

Pretty print a proof tree (in ASCII form).

proofNodesWithDepth :: Proof -> [(Int, String)]

Extract proof nodes with corresponding depths.

ppNodesWithDepth :: [(Int, String)] -> String

Pretty-print depth-annotated nodes as a ASCII tree-like structure.

>>> putStr $ ppNodesWithDepth (zip [0,1,2,1,2,3,3,2,3] (map show [0..]))
0
├─ 1
│  └─ 2
└─ 3
   ├─ 4
   │  ├─ 5
   │  └─ 6
   └─ 7
      └─ 8

depthsTree :: [Int] -> [String]

Pretty-print an ASCII tree-like structure based on depth list.

>>> mapM_ putStrLn $ depthsTree [0,1,2,1,2,3,3,2,3]
└─
   ├─
   │  └─
   └─
      ├─
      │  ├─
      │  └─
      └─
         └─

Provers

proveWithDFS

Arguments

:: Int
Maximum search depth.

-> Rules
Set of rules.

-> Sequent
Starting sequent (one to prove).

-> Logic (ProofTree Sequent)
A proof tree, possibly with unproven sequents in leaves (for when maximum search depth is insufficient).

A simple DFS proof search algorithm for a given set of rules.

proveWithBFSviaDFS

Arguments

:: Int
Maximum number of limited DFS iterations.

-> Int
Maximum depth per one DFS iteration.

-> Rules
A set of rules.

-> Sequent
A starting sequent (one to prove).

-> Logic Proof

A simple k-depth proof search algorithm for a given set of rules.

proveWithDFS'

Arguments

:: Int
Maximum search depth.

-> Rules
Set of rules.

-> Sequent
Starting sequent (one to prove).

-> Maybe Proof

Search for a proof using simple DFS algorithm (see proveWithDFS).

proveWithBFSviaDFS'

Arguments

:: Int
Maximum number of limited DFS iterations.

-> Int
Maximum depth per one DFS iteration.

-> Rules
A set of rules.

-> Sequent
A starting sequent (one to prove).

-> Maybe Proof

Search for a proof using a combination of BFS and DFS algorithms (see proveWithBFS).

With printers

proveAndPrintBFS

Arguments

:: Int
Maximum search depth.

-> Rules
Set of rules.

-> Sequent
Starting sequent (one to prove).

-> IO ()

Search for a proof using BFS and print proof tree (see ppProof).

proveAndPrintBFSviaDFS :: Int -> Int -> Rules -> Sequent -> IO ()

Search for a proof using a combination of BFS and DFS and print proof tree (see ppProof).

proveAndPrintDFS :: Int -> Rules -> Sequent -> IO ()

Search for a proof using DFS and print proof tree (see ppProof).

Examples

ex1 :: Sequent

>>> putStrLn (ppSequent ex1)
⋅ | (φ ∨ ζ) ∧ (ψ ∨ χ) ⊢ φ ∧ ψ ∨ ζ ∧ ψ ∨ χ

>>> proveAndPrintBFS 8 rulesLJ ex1
[∧L]  ⋅ | (φ ∨ ζ) ∧ (ψ ∨ χ) ⊢ φ ∧ ψ ∨ ζ ∧ ψ ∨ χ
└─ [∨L]  ⋅ | φ ∨ ζ, ψ ∨ χ ⊢ φ ∧ ψ ∨ ζ ∧ ψ ∨ χ
   ├─ [∨L]  ⋅ | φ, ψ ∨ χ ⊢ φ ∧ ψ ∨ ζ ∧ ψ ∨ χ
   │  ├─ [∨R₁]  ⋅ | ψ, φ ⊢ φ ∧ ψ ∨ ζ ∧ ψ ∨ χ
   │  │  └─ [∧R]  ⋅ | ψ, φ ⊢ φ ∧ ψ
   │  │     ├─ [Ax]  ⋅ | ψ, φ ⊢ φ
   │  │     └─ [Ax]  ⋅ | ψ, φ ⊢ ψ
   │  └─ [∨R₂]  ⋅ | χ, φ ⊢ φ ∧ ψ ∨ ζ ∧ ψ ∨ χ
   │     └─ [∨R₂]  ⋅ | χ, φ ⊢ ζ ∧ ψ ∨ χ
   │        └─ [Ax]  ⋅ | χ, φ ⊢ χ
   └─ [∨L]  ⋅ | ζ, ψ ∨ χ ⊢ φ ∧ ψ ∨ ζ ∧ ψ ∨ χ
      ├─ [∨R₂]  ⋅ | ψ, ζ ⊢ φ ∧ ψ ∨ ζ ∧ ψ ∨ χ
      │  └─ [∨R₁]  ⋅ | ψ, ζ ⊢ ζ ∧ ψ ∨ χ
      │     └─ [∧R]  ⋅ | ψ, ζ ⊢ ζ ∧ ψ
      │        ├─ [Ax]  ⋅ | ψ, ζ ⊢ ζ
      │        └─ [Ax]  ⋅ | ψ, ζ ⊢ ψ
      └─ [∨R₂]  ⋅ | χ, ζ ⊢ φ ∧ ψ ∨ ζ ∧ ψ ∨ χ
         └─ [∨R₂]  ⋅ | χ, ζ ⊢ ζ ∧ ψ ∨ χ
            └─ [Ax]  ⋅ | χ, ζ ⊢ χ

ex2 :: Sequent

>>> putStrLn (ppSequent ex2)
⋅ | φ ∧ ψ ∨ ζ ∧ χ ⊢ (φ ∨ ζ) ∧ (ψ ∨ χ)

>>> proveAndPrintBFS 8 rulesLJ ex2
[∧R]  ⋅ | φ ∧ ψ ∨ ζ ∧ χ ⊢ (φ ∨ ζ) ∧ (ψ ∨ χ)
├─ [∨L]  ⋅ | φ ∧ ψ ∨ ζ ∧ χ ⊢ φ ∨ ζ
│  ├─ [∧L]  ⋅ | φ ∧ ψ ⊢ φ ∨ ζ
│  │  └─ [∨R₁]  ⋅ | φ, ψ ⊢ φ ∨ ζ
│  │     └─ [Ax]  ⋅ | φ, ψ ⊢ φ
│  └─ [∧L]  ⋅ | ζ ∧ χ ⊢ φ ∨ ζ
│     └─ [∨R₂]  ⋅ | ζ, χ ⊢ φ ∨ ζ
│        └─ [Ax]  ⋅ | ζ, χ ⊢ ζ
└─ [∨L]  ⋅ | φ ∧ ψ ∨ ζ ∧ χ ⊢ ψ ∨ χ
   ├─ [∧L]  ⋅ | φ ∧ ψ ⊢ ψ ∨ χ
   │  └─ [∨R₁]  ⋅ | φ, ψ ⊢ ψ ∨ χ
   │     └─ [Ax]  ⋅ | φ, ψ ⊢ ψ
   └─ [∧L]  ⋅ | ζ ∧ χ ⊢ ψ ∨ χ
      └─ [∨R₂]  ⋅ | ζ, χ ⊢ ψ ∨ χ
         └─ [Ax]  ⋅ | ζ, χ ⊢ χ

ex3 :: Sequent

>>> putStrLn (ppSequent ex3)
⋅ | x ≡ y ∧ y ≡ z ⊢ z ≡ x

>>> proveAndPrintBFS 35 (rulesLJE <> rulesLEQ) ex3
[∧L]  ⋅ | x ≡ y ∧ y ≡ z ⊢ z ≡ x
└─ [≡L(trans)]  ⋅ | x ≡ y, y ≡ z ⊢ z ≡ x
   └─ [≡L(sym)]  ⋅ | x ≡ z, x ≡ y, y ≡ z ⊢ z ≡ x
      └─ [Ax]  ⋅ | z ≡ x, x ≡ z, x ≡ y, y ≡ z ⊢ z ≡ x

ex4 :: Sequent

>>> putStrLn (ppSequent ex4)
⋅ | (t ≡ s ∨ t ≡ u) ∧ s ≡ u ⊢ t ≡ s

>>> proveAndPrintBFS 8 rulesLJE ex4
[∧L]  ⋅ | (t ≡ s ∨ t ≡ u) ∧ s ≡ u ⊢ t ≡ s
└─ [∨L]  ⋅ | t ≡ s ∨ t ≡ u, s ≡ u ⊢ t ≡ s
   ├─ [Ax]  ⋅ | t ≡ s, s ≡ u ⊢ t ≡ s
   └─ [≡L(sym)]  ⋅ | t ≡ u, s ≡ u ⊢ t ≡ s
      └─ [≡L(trans)]  ⋅ | u ≡ s, t ≡ u, s ≡ u ⊢ t ≡ s
         └─ [Ax]  ⋅ | t ≡ s, u ≡ s, t ≡ u, s ≡ u ⊢ t ≡ s

ex5 :: Sequent

>>> putStrLn (ppSequent ex5)
⋅ | (≤(t, s) ∨ ≤(s, u)) ∧ ≤(t, u) ⊢ ≤(t, u) ∨ ≤(s, t) ∨ ≤(u, s)

>>> proveAndPrintBFS 5 (rulesLJE <> rulesLEQ) ex5
[∧L]  ⋅ | (≤(t, s) ∨ ≤(s, u)) ∧ ≤(t, u) ⊢ ≤(t, u) ∨ ≤(s, t) ∨ ≤(u, s)
└─ [∨R₁]  ⋅ | ≤(t, s) ∨ ≤(s, u), ≤(t, u) ⊢ ≤(t, u) ∨ ≤(s, t) ∨ ≤(u, s)
   └─ [∨R₁]  ⋅ | ≤(t, s) ∨ ≤(s, u), ≤(t, u) ⊢ ≤(t, u) ∨ ≤(s, t)
      └─ [Ax]  ⋅ | ≤(t, s) ∨ ≤(s, u), ≤(t, u) ⊢ ≤(t, u)